git-svn-id: file:///Users/will/rendale@51 651630e0-a32e-0410-a001-f4242d931729

+++ Rails/notable/Trunk/app/models/note_file.rb
@@ -2,14 +2,25 @@ class NoteFile < ActiveRecord::Base
 	belongs_to :note
 	before_destroy :remove_file  
 
+  attr_reader :error_message
+
 	def add_file(file)
+	  begin  
+	    if file.length > 5.megabytes
+        @error_message = "File exceeds maximum size of 5 megabytes"
+        return false
+	    end
   	  self.filename = sanitize_filename(file.original_filename)
       self.systempath = self.note.user.login + "/" + self.note.id.to_s + "/"
 	    self.filetype = file.content_type	    
 	    
 	    FileUtils.mkdir_p(RAILS_ROOT + "/files/" + self.systempath)
+      if File.exists?(RAILS_ROOT + "/files/" + self.systempath + self.filename)
+        @error_message = "A file with that name already exists."
+        return false
+      else
 	    File.open(RAILS_ROOT + "/files/" + self.systempath + self.filename, "wb") { |f| f.write(file.read) }
-
+      end
       if file.content_type=~/image/
         img = Magick::Image::read(RAILS_ROOT + "/files/" + self.systempath + self.filename).first        
 			  yxratio = (img.rows.to_f/img.columns.to_f) * 32
@@ -18,9 +29,15 @@ class NoteFile < ActiveRecord::Base
   			begin
 			    img.crop!(Magick::CenterGravity,32,32)
         rescue 
+          @error_message = "Error encountered creating thumbnail"
+          return false
         end
         img.write RAILS_ROOT + "/files/" + self.systempath + "thumb_" + self.filename
-      end  
+      end
+    rescue
+      @error_message = "Error processing file"
+      return false
+    end  
     return true
   end
